ipatests: fix circular import for collect_logs
Move collect_logs function from util to avoid a circular import.
